Migrate postcss-focus-within
