--- plugins/postcss-focus-within/.gitattributes
@@ -1,1 +1,0 @@
-* text=auto eol=lf
