--- plugins/postcss-focus-within/.github/workflows/test.yml
@@ -1,18 +1,0 @@
-name: test
-on:
-  push:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        node: [12, 16]
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v2
-        with:
-          node-version: ${{ matrix.node }}
-
-      - run: yarn install --ignore-scripts
-      - run: yarn run test
