+++ plugins/postcss-focus-within/INSTALL.md
@@ -165,6 +165,6 @@ grunt.initConfig({
 [PostCSS]: https://github.com/postcss/postcss
 [PostCSS CLI]: https://github.com/postcss/postcss-cli
 [PostCSS Loader]: https://github.com/postcss/postcss-loader
-[PostCSS]: https://github.com/csstools/postcss-focus-within
+[PostCSS]: https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-focus-within
 [React App Rewire PostCSS]: https://github.com/csstools/react-app-rewire-postcss
 [React App Rewired]: https://github.com/timarney/react-app-rewired
