+++ plugins/postcss-focus-within/test/basic.replacewith.expect.css
@@ -6,8 +6,7 @@
 	order: 1;
 }
 
-.focus-within,
-.focus-within test,
+.focus-within,.focus-within test,
 test .focus-within,
 test test.focus-within,
 test .focus-within test,
